weather icon update

+++ lib/modules/weather_home/model/weather_temperature.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 class WeatherTemperature {
-  final double actualTemp;
+  final int actualTemp;
   final double feelsLike;
   final double minTemp;
   final double maxTemp;
@@ -25,7 +25,7 @@ class WeatherTemperature {
 
   factory WeatherTemperature.fromMap(Map<String, dynamic> map) {
     return WeatherTemperature(
-      actualTemp: map['actual'] as double,
+      actualTemp: map['actual'].toInt(),
       feelsLike: map['feelsLike'] as double,
       minTemp: map['min'] as double,
       maxTemp: map['max'] as double,
